feat: display end page drawer, refactor slightly to shard out motionbox container component
