+++ frontend/src/features/admin-form/create/builder-and-design/BuilderAndDesignDrawer/BuilderAndDesignDrawer.tsx
@@ -1,15 +1,10 @@
 import { useMemo } from 'react'
-import { Flex } from '@chakra-ui/react'
-import { AnimatePresence } from 'framer-motion'
 
-import { useIsMobile } from '~hooks/useIsMobile'
-import { MotionBox } from '~components/motion'
-
+import { CreatePageDrawerContainer } from '../../common/CreatePageDrawer/CreatePageDrawerContainer'
 import {
   DrawerTabs,
   useCreatePageSidebar,
 } from '../../common/CreatePageSidebarContext'
-import EndPageDrawer from '../../end-page/EndPageDrawer'
 import {
   FieldBuilderState,
   fieldBuilderStateSelector,
@@ -21,30 +16,8 @@ import { EditFieldDrawer } from './EditFieldDrawer'
 import { FieldListDrawer } from './FieldListDrawer'
 
 export const BuilderAndDesignDrawer = (): JSX.Element | null => {
-  const isMobile = useIsMobile()
-  const { activeTab, isDrawerOpen, drawerRef } = useCreatePageSidebar()
+  const { activeTab } = useCreatePageSidebar()
   const fieldBuilderState = useFieldBuilderStore(fieldBuilderStateSelector)
-
-  const drawerMotionProps = useMemo(() => {
-    return {
-      initial: { width: 0 },
-      animate: {
-        maxWidth: isMobile ? '100%' : '33.25rem',
-        width: isMobile ? '100%' : '36%',
-        transition: {
-          bounce: 0,
-          duration: 0.2,
-        },
-      },
-      exit: {
-        width: 0,
-        opacity: 0,
-        transition: {
-          duration: 0.2,
-        },
-      },
-    }
-  }, [isMobile])
 
   const renderDrawerContent: JSX.Element | null = useMemo(() => {
     switch (activeTab) {
@@ -60,27 +33,12 @@ export const BuilderAndDesignDrawer = (): JSX.Element | null => {
       case DrawerTabs.Design:
         return <DesignDrawer />
       default:
-        // Logic or endpage drawer open
-        return <EndPageDrawer />
+        // Logic or endpage open
+        return null
     }
   }, [fieldBuilderState, activeTab])
 
   return (
-    <AnimatePresence>
-      {isDrawerOpen ? (
-        <MotionBox
-          bg="white"
-          key="sidebar"
-          pos="relative"
-          as="aside"
-          overflow="hidden"
-          {...drawerMotionProps}
-        >
-          <Flex w="100%" h="100%" flexDir="column" ref={drawerRef}>
-            {renderDrawerContent}
-          </Flex>
-        </MotionBox>
-      ) : null}
-    </AnimatePresence>
+    <CreatePageDrawerContainer>{renderDrawerContent}</CreatePageDrawerContainer>
   )
 }
