+++ frontend/src/features/admin-form/create/builder-and-design/BuilderAndDesignDrawer/DesignDrawer/DesignDrawer.tsx
@@ -14,7 +14,6 @@ import {
   Flex,
   FormControl,
   FormLabel,
-  Tabs,
   Text,
   Textarea,
 } from '@chakra-ui/react'
@@ -45,13 +44,13 @@ import {
   startPageDataSelector,
   stateSelector,
   useDesignStore,
-} from '../../useDesignStore'
+} from '../../../builder-and-design/useDesignStore'
 import {
   setIsDirtySelector,
   useDirtyFieldStore,
-} from '../../useDirtyFieldStore'
-import { validateNumberInput } from '../../utils/validateNumberInput'
-import { CreatePageDrawerCloseButton } from '../CreatePageDrawerCloseButton'
+} from '../../../builder-and-design/useDirtyFieldStore'
+import { validateNumberInput } from '../../../builder-and-design/utils/validateNumberInput'
+import { CreatePageDrawerCloseButton } from '../../../common/CreatePageDrawer/CreatePageDrawerCloseButton'
 import { DrawerContentContainer } from '../EditFieldDrawer/edit-fieldtype/common/DrawerContentContainer'
 import { FormFieldDrawerActions } from '../EditFieldDrawer/edit-fieldtype/common/FormFieldDrawerActions'
 import {
@@ -432,7 +431,7 @@ export const DesignDrawer = ({
   if (!startPageData) return null
 
   return (
-    <Tabs pos="relative" h="100%" display="flex" flexDir="column">
+    <Flex pos="relative" h="100%" display="flex" flexDir="column">
       <Box pt="1rem" px="1.5rem" bg="white">
         <Flex justify="space-between">
           <Text textStyle="subhead-3" color="secondary.500" mb="1rem">
@@ -443,6 +442,6 @@ export const DesignDrawer = ({
         <Divider w="auto" mx="-1.5rem" />
       </Box>
       <DesignInput />
-    </Tabs>
+    </Flex>
   )
 }
