+++ frontend/src/features/admin-form/create/common/CreatePageDrawer/index.ts
@@ -1,0 +1,2 @@
+export * from './CreatePageDrawerCloseButton'
+export * from './CreatePageDrawerContainer'
