+++ frontend/src/features/admin-form/create/common/index.ts
@@ -1,3 +1,4 @@
 export * from './CreatePageContent'
+export * from './CreatePageDrawer'
 export * from './CreatePageSidebar'
 export * from './CreatePageSidebarContext'
