--- frontend/src/features/admin-form/create/end-page/EndPageDrawer/index.ts
@@ -1,1 +1,0 @@
-export { EndPageDrawer as default } from './EndPageDrawer'
